Introduce useEmblaCarousel() :tropical_fish:
