+++ src/utils.ts
@@ -1,0 +1,1 @@
+export const canUseDOM = !!(typeof window !== 'undefined' && window.document)
